fix dev review issues
