+++ modules/gltf/src/index.ts
@@ -49,7 +49,9 @@ export type {
 
 export type {
   GLTF_EXT_structural_metadata_GLTF,
-  GLTF_EXT_structural_metadata_Schema
+  GLTF_EXT_structural_metadata_Schema,
+  GLTF_EXT_structural_metadata_PropertyTable,
+  GLTF_EXT_structural_metadata_PropertyTexture
 } from './lib/types/gltf-ext-structural-metadata-schema';
 
 export type {
@@ -60,8 +62,6 @@ export type {
 export {name as EXT_MESH_FEATURES} from './lib/extensions/EXT_mesh_features';
 export {name as EXT_STRUCTURAL_METADATA} from './lib/extensions/EXT_structural_metadata';
 export {name as EXT_FEATURE_METADATA} from './lib/extensions/deprecated/EXT_feature_metadata';
-export {getPropertyTableFromExtFeatureMetadata} from './lib/extensions/deprecated/EXT_feature_metadata';
-export {getPropertyTableFromExtStructuralMetadata} from './lib/extensions/EXT_structural_metadata';
 
 // Postprocessed types (modified GLTF types)
 export type {
